+++ lib/puma/configuration.rb
@@ -137,6 +137,8 @@ module Puma
       @file_dsl    = DSL.new(@options.file_options, self)
       @default_dsl = DSL.new(@options.default_options, self)
 
+      default_options[:preload_app] = (default_options[:workers] > 1 && Puma::Plugin.new.workers_supported?)
+
       if block
         configure(&block)
       end
@@ -167,14 +169,19 @@ module Puma
       self
     end
 
+    def default_max_threads
+      return 5 if Puma.mri?
+      16
+    end
+
     def puma_default_options
       {
-        :min_threads => 0,
-        :max_threads => 16,
+        :min_threads => Integer(ENV['RAILS_MIN_THREADS'] || ENV['MIN_THREADS'] || 0),
+        :max_threads => Integer(ENV['RAILS_MAX_THREADS'] || ENV['MAX_THREADS'] || default_max_threads),
         :log_requests => false,
         :debug => false,
         :binds => ["tcp://#{DefaultTCPHost}:#{DefaultTCPPort}"],
-        :workers => 0,
+        :workers => Integer(ENV['WEB_CONCURRENCY'] || 0),
         :daemon => false,
         :mode => :http,
         :worker_timeout => DefaultWorkerTimeout,